refactor(P2-4): TaskStatus 重複定義を解消し types/wbs の union に一本化
- 本番の残存 Prisma enum 依存を除去: infrastructures/evm/wbs-evm-repository.ts の
  TaskStatus を @prisma/client → @/types/wbs(union) へ（型注釈・キャストのみで挙動不変）
- evm テスト4件も @/types union へ統一（evm-service.test.ts はランタイム参照
  TaskStatus.X を文字列リテラル 'X' に変換）
- ④ ファイル名タイポ修正(P4-1): value-object/project-status.ts(中身は TaskStatus クラス)
  を task-status.ts へリネームし import 20 箇所を更新
- 本番型チェック(tsc)エラーなし、関連ユニットテスト 186 件 green
