+++ src/__integration_tests__/assignee-gantt/assignee-gantt.integration.test.ts
@@ -7,7 +7,7 @@ import { ICompanyHolidayRepository } from '@/applications/calendar/icompany-holi
 import { Task } from '@/domains/task/task';
 import { WbsAssignee } from '@/domains/wbs/wbs-assignee';
 import { TaskNo } from '@/domains/task/value-object/task-id';
-import { TaskStatus } from '@/domains/task/value-object/project-status';
+import { TaskStatus } from '@/domains/task/value-object/task-status';
 import { Period } from '@/domains/task/period';
 import { CompanyHoliday } from '@/domains/calendar/company-calendar';
 import { PeriodType } from '@/domains/task/value-object/period-type';
