+++ src/__integration_tests__/helpers.ts
@@ -4,7 +4,7 @@ import { Project } from '@/domains/project/project';
 import { Phase } from '@/domains/phase/phase';
 import { PhaseCode } from '@/domains/phase/phase-code';
 import { Task } from '@/domains/task/task';
-import { TaskStatus } from '@/domains/task/value-object/project-status';
+import { TaskStatus } from '@/domains/task/value-object/task-status';
 import { TaskNo } from '@/domains/task/value-object/task-id';
 import { Period } from '@/domains/task/period';
 import { PeriodType } from '@/domains/task/value-object/period-type';
