+++ src/__tests__/applications/task-factory.test.ts
@@ -6,7 +6,7 @@ import { Phase } from "@/domains/phase/phase";
 import { PhaseCode } from "@/domains/phase/phase-code";
 import { Task } from "@/domains/task/task";
 import { TaskNo } from "@/domains/task/value-object/task-id";
-import { TaskStatus } from "@/domains/task/value-object/project-status";
+import { TaskStatus } from "@/domains/task/value-object/task-status";
 
 // モックの設定
 jest.mock("@/applications/task/itask-repository");
