+++ src/__tests__/applications/task/task-factory.test.ts
@@ -5,7 +5,7 @@ import { TaskNo } from "@/domains/task/value-object/task-id";
 import { Task } from "@/domains/task/task";
 import { Phase } from "@/domains/phase/phase";
 import { PhaseCode } from "@/domains/phase/phase-code";
-import { TaskStatus } from "@/domains/task/value-object/project-status";
+import { TaskStatus } from "@/domains/task/value-object/task-status";
 
 // モッククラス
 class MockTaskRepository implements ITaskRepository {
