+++ src/__tests__/dashboard/dashboard-application-service.test.ts
@@ -5,7 +5,7 @@ import { ITaskRepository } from "@/applications/task/itask-repository";
 import { Project } from "@/domains/project/project";
 import { ProjectStatus } from "@/domains/project/project-status";
 import { Task } from "@/domains/task/task";
-import { TaskStatus } from "@/domains/task/value-object/project-status";
+import { TaskStatus } from "@/domains/task/value-object/task-status";
 import { TaskNo } from "@/domains/task/value-object/task-id";
 import { Wbs } from "@/domains/wbs/wbs";
