+++ src/__tests__/domains/assignee-workload/workload-calculation-service.test.ts
@@ -1,7 +1,7 @@
 import { WorkloadCalculationService } from '@/domains/assignee-workload/workload-calculation.service';
 import { Task } from '@/domains/task/task';
 import { TaskNo } from '@/domains/task/value-object/task-id';
-import { TaskStatus } from '@/domains/task/value-object/project-status';
+import { TaskStatus } from '@/domains/task/value-object/task-status';
 import { Period } from '@/domains/task/period';
 import { PeriodType } from '@/domains/task/value-object/period-type';
 import { ManHour } from '@/domains/task/man-hour';
