+++ src/__tests__/domains/sync/WbsSyncService.test.ts
@@ -15,7 +15,7 @@
 // import { ITaskRepository } from '@/applications/task/itask-repository';
 // import { Task } from '@/domains/task/task';
 // import { TaskNo } from '@/domains/task/value-object/task-id';
-// import { TaskStatus } from '@/domains/task/value-object/project-status';
+// import { TaskStatus } from '@/domains/task/value-object/task-status';
 
 describe('WbsSyncService', () => {
   describe('placeholder', () => {
