+++ src/infrastructures/evm/wbs-evm-repository.ts
@@ -12,7 +12,7 @@ import {
 import type { IWbsQueryRepository } from '@/applications/wbs/query/wbs-query-repository';
 import { TaskEvmData } from '@/domains/evm/task-evm-data';
 import { EvmCalculationMode } from '@/domains/evm/evm-metrics';
-import { TaskStatus } from '@prisma/client';
+import { TaskStatus } from '@/types/wbs';
 
 @injectable()
 export class WbsEvmRepository implements IWbsEvmRepository {
